* Fix compilation warning in Timeval class when compiling in Windows * Call StopInquiry() when an inquiry timeout occurs with WIDCOMM Bluetooth stack. Hopefully, this will reset the inquiry and the BT stack so that one can inquiry again the next scan interval.

+++ src/libcpphaggle/Timeval.cpp
@@ -74,8 +74,8 @@ Timeval::Timeval(const string str)
 	
 Timeval::Timeval(const double _t)
 {
-	t.tv_sec = _t;
-	t.tv_usec = (_t - t.tv_sec) * 1000000.0;
+	t.tv_sec = (long)_t;
+	t.tv_usec = (long)((_t - t.tv_sec) * 1000000.0);
 }
 
 Timeval& Timeval::setNow()
